# docs/swagger.yaml
basePath: /
definitions:
  handlers.CollectionItemRequest:
    properties:
      collection_id:
        type: integer
      collection_item:
        type: string
      relation_collection_items_id:
        type: integer
    type: object
  handlers.CollectionRequest:
    properties:
      collection_name:
        type: string
    required:
    - collection_name
    type: object
...
  structs.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Kora API server for form management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kora API
  version: "1.0"
paths:
  /collection_items:
    post:
      consumes:
      - application/json
      description: Create a new collection item
      parameters:
      - description: Collection Item Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CollectionItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Create collection item
      tags:
      - collection-items
  /collection_items/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a collection item by its ID
      parameters:
      - description: Collection Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Delete collection item
      tags:
      - collection-items
    get:
      consumes:
      - application/json
      description: Retrieve a collection item by its ID
      parameters:
      - description: Collection Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get collection item
      tags:
      - collection-items
    put:
      consumes:
      - application/json
      description: Update an existing collection item by its ID
      parameters:
      - description: Collection Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Collection Item Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CollectionItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Update collection item
      tags:
      - collection-items
  /collections:
    get:
      consumes:
      - application/json
      description: Retrieve all collections
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get all collections
      tags:
      - collections
    post:
      consumes:
      - application/json
      description: Create a new collection with the provided name
      parameters:
      - description: Collection Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CollectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Create a new collection
      tags:
      - collections
  /collections/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a collection by its ID
      parameters:
      - description: Collection ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Delete collection
      tags:
      - collections
    get:
      consumes:
      - application/json
      description: Retrieve a collection by its ID
      parameters:
      - description: Collection ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Get collection by ID
      tags:
      - collections
    put:
      consumes:
      - application/json
      description: Update an existing collection by its ID
      parameters:
      - description: Collection ID
        in: path
        name: id
        required: true
        type: integer
      - description: Collection Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CollectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      summary: Update collection
      tags:
      - collections
...
schemes:
- http
- https
swagger: "2.0"